Fix styling and update state on LikeButton

+++ src/components/LikeButton.js
@@ -1,21 +1,17 @@
-// src/components/LikeButton.js
 import React from "react";
 import axios from "axios";
 import styles from "../styles/LikeButton.module.css";
 
-const LikeButton = ({ review, setReviews }) => {
-
+const LikeButton = ({ review, updateReview }) => {
   const handleLike = async (reviewId) => {
     try {
       const { data } = await axios.post("/likes/", { review: reviewId });
 
-      setReviews((prevReviews) =>
-        prevReviews.map((review) =>
-          review.id === reviewId
-            ? { ...review, likes_count: review.likes_count + 1, like_id: data.id }
-            : review
-        )
-      );
+      updateReview({
+        ...review,
+        likes_count: review.likes_count + 1,
+        like_id: data.id,
+      });
     } catch (err) {
       console.error("Error liking review:", err);
     }
@@ -23,15 +19,18 @@ const LikeButton = ({ review, setReviews }) => {
 
   const handleUnlike = async (reviewId, likeId) => {
     try {
+      if (!likeId) {
+        console.error("No like_id available for unliking.");
+        return;
+      }
+
       await axios.delete(`/likes/${likeId}`);
 
-      setReviews((prevReviews) =>
-        prevReviews.map((review) =>
-          review.id === reviewId
-            ? { ...review, likes_count: review.likes_count - 1, like_id: null }
-            : review
-        )
-      );
+      updateReview({
+        ...review,
+        likes_count: review.likes_count - 1,
+        like_id: null,
+      });
     } catch (err) {
       console.error("Error unliking review:", err.response);
     }
@@ -40,7 +39,10 @@ const LikeButton = ({ review, setReviews }) => {
   return (
     <span className="text-end">
       {review.like_id ? (
-        <span className="text-success me-2" onClick={() => handleUnlike(review.id, review.like_id)}>
+        <span
+          className="text-success me-2"
+          onClick={() => handleUnlike(review.id, review.like_id)}
+        >
           <i className={`fa-solid fa-thumbs-up ${styles.likedHeart}`} />
         </span>
       ) : (
